requirements.txt
Django==4.0.4
djangorestframework==3.13.1
celery==5.2.6